requirements.txt
langchain>=0.3.0
langchain-community>=0.0.21
langchain-ollama>=0.1.0
ollama
requests
python-dotenv
fastapi
uvicorn
